requirements.txt
altgraph==0.17.4
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
lxml==6.0.1
packaging==25.0
pillow==10.4.0
psutil==7.1.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
pyperclip==1.9.0
requests==2.32.5
setuptools==80.9.0
soupsieve==2.8
speedtest-cli==2.1.3
ttkbootstrap==1.14.2
typing_extensions==4.15.0
urllib3==2.5.0
yt-dlp==2025.9.5
